requirements.txt
beautifulsoup4==4.8.0
cachetools==3.1.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cssselect2==0.2.2
defusedxml==0.6.0
Django==2.2.5
django-appconf==1.0.3
django-bootstrap4==1.0.1
django-extensions==2.2.3
django-filter==2.2.0
django-googledrive-storage==1.3.5
django-summernote==0.8.11.4
django-weasyprint==0.5.4
djangorestframework==3.10.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
gunicorn==19.9.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
oauthlib==3.1.0
Pillow==6.2.0
psycopg2==2.8.3
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pydot==1.4.1
pyflakes==2.1.1
pyparsing==2.4.2
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
tinycss2==1.0.2
uritemplate==3.0.0
urllib3==1.25.6
WeasyPrint==50
webencodings==0.5.1
whitenoise==4.1.4
